+++ vite.config.js
@@ -174,7 +174,7 @@ export default defineConfig({
         },
     },
     server: {
-        // cors: true,
+        cors: true,
 
         // BrowserStackLocal
         allowedHosts: true,